IMPALA-6177: Cleanup incomplete handcrafted IRs before finalizing module
Currently, if an error is encountered during the creation of a
handcrafted codegen method, then the resulting IR is left in an
incomplete state. This patch ensures that all such IRs are cleaned up
(method body is truncated) before the llvm module is finalized.

Testing:
Tested manually by executing the following query and inspecting
the dump of the optimized llvm module.

Change-Id: If975cfb3906482b36dd6ede32ca81de6fcee1d7f